+++ src/components/games/index.ts
@@ -1,3 +1,4 @@
 export { default as TriviaGame } from './TriviaGame';
 export { default as DevQuest } from './DevQuest';
 export { default as CodeBotBattler } from './CodeBotBattler';
+export { default as DevType } from './DevType';